requirements.txt
docx==0.2.4
Pillow==9.3.0
python_docx==0.8.11
streamlit==1.14.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
